+++ serviYa/src/routes/misTurnos/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import type { PageData } from './$types';
+
+	export let data: PageData;
+</script>
